agency-swarm-cursor-agent/presentation_agency_1/requirements.txt
agency-swarm==0.4.1
python-pptx==1.0.2
PyPDF2==3.0.1
matplotlib==3.9.3
python-dotenv==1.0.1
openai==1.57.3 